+++ src/components/v1/V1Project/TokensSection.tsx
@@ -9,7 +9,7 @@ import ParticipantsModal from 'components/modals/ParticipantsModal'
 import SectionHeader from 'components/SectionHeader'
 import { ProjectMetadataContext } from 'contexts/projectMetadataContext'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
-import useERC20BalanceOf from 'hooks/ERC20BalanceOf'
+import useERC20BalanceOf from 'hooks/ERC20/ERC20BalanceOf'
 import useCanPrintPreminedTokens from 'hooks/v1/contractReader/CanPrintPreminedTokens'
 import useReservedTokensOfProject from 'hooks/v1/contractReader/ReservedTokensOfProject'
 import useTotalBalanceOf from 'hooks/v1/contractReader/TotalBalanceOf'
