+++ src/components/v2v3/V2V3Project/V2V3ManageTokensSection/AccountBalanceDescription/AccountBalanceDescription.tsx
@@ -7,7 +7,7 @@ import { FEATURE_FLAGS } from 'constants/featureFlags'
 import { ProjectMetadataContext } from 'contexts/projectMetadataContext'
 import { V2V3ProjectContext } from 'contexts/v2v3/V2V3ProjectContext'
 import { VeNftContext } from 'contexts/veNftContext'
-import useERC20BalanceOf from 'hooks/ERC20BalanceOf'
+import useERC20BalanceOf from 'hooks/ERC20/ERC20BalanceOf'
 import useTotalBalanceOf from 'hooks/v2v3/contractReader/TotalBalanceOf'
 import useUserUnclaimedTokenBalance from 'hooks/v2v3/contractReader/UserUnclaimedTokenBalance'
 import { useV2ConnectedWalletHasPermission } from 'hooks/v2v3/contractReader/V2ConnectedWalletHasPermission'
