+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ReconfigureFundingCycleSettingsPage/NftSummarySection.tsx
@@ -4,7 +4,7 @@ import ExternalLink from 'components/ExternalLink'
 import Paragraph from 'components/Paragraph'
 import { NFT_IMAGE_SIDE_LENGTH } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/NftUpload'
 import { useAppSelector } from 'redux/hooks/AppSelector'
-import { DEFAULT_NFT_MAX_SUPPLY } from 'hooks/NftRewards'
+import { DEFAULT_NFT_MAX_SUPPLY } from 'hooks/JB721Delegate/NftRewards'
 import { classNames } from 'utils/classNames'
 
 export default function NftSummarySection() {
