+++ src/components/v2v3/V2V3Project/modals/V2V3ProjectTokenBalancesModal/V2V3ProjectTokenBalance.tsx
@@ -1,7 +1,7 @@
 import { Trans } from '@lingui/macro'
 import V2V3ProjectHandleLink from 'components/v2v3/shared/V2V3ProjectHandleLink'
 import { V2V3ProjectContext } from 'contexts/v2v3/V2V3ProjectContext'
-import useSymbolOfERC20 from 'hooks/SymbolOfERC20'
+import useSymbolOfERC20 from 'hooks/ERC20/SymbolOfERC20'
 import useProjectHandle from 'hooks/v2v3/contractReader/ProjectHandle'
 import useProjectToken from 'hooks/v2v3/contractReader/ProjectToken'
 import useTotalBalanceOf from 'hooks/v2v3/contractReader/TotalBalanceOf'
