+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/NftCollectionDetailsForm.tsx
@@ -3,7 +3,7 @@ import { Form, FormInstance, FormProps, Input } from 'antd'
 import Loading from 'components/Loading'
 import { NftRewardsContext } from 'contexts/nftRewardsContext'
 import { useAppSelector } from 'redux/hooks/AppSelector'
-import { useNftCollectionMetadata } from 'hooks/NftCollectionMetadata'
+import { useNftCollectionMetadata } from 'hooks/JB721Delegate/NftCollectionMetadata'
 import { useCallback, useContext, useEffect, useMemo } from 'react'
 import {
   defaultNftCollectionDescription,
