+++ src/components/veNft/VeNftStakingForm.tsx
@@ -5,8 +5,8 @@ import { Button, Col, Form, Row, Space } from 'antd'
 import { useForm, useWatch } from 'antd/lib/form/Form'
 import { useContext, useEffect, useMemo, useState } from 'react'
 import { V2V3ProjectContext } from 'contexts/v2v3/V2V3ProjectContext'
-import useERC20Allowance from 'hooks/ERC20Allowance'
-import useERC20BalanceOf from 'hooks/ERC20BalanceOf'
+import useERC20Allowance from 'hooks/ERC20/ERC20Allowance'
+import useERC20BalanceOf from 'hooks/ERC20/ERC20BalanceOf'
 import useERC20Approve from 'hooks/veNft/transactor/ERC20ApproveTx'
 import CustomBeneficiaryInput from 'components/veNft/formControls/CustomBeneficiaryInput'
 import LockDurationSelectInput from 'components/veNft/formControls/LockDurationSelectInput'
