+++ src/hooks/ERC20/ERC20BalanceOf.ts
@@ -1,8 +1,8 @@
 import { BigNumber } from '@ethersproject/bignumber'
-import { useErc20Contract } from 'hooks/Erc20Contract'
+import { useErc20Contract } from 'hooks/ERC20/Erc20Contract'
 import { bigNumbersDiff } from 'utils/bigNumbers'
 
-import useContractReader from './v2v3/contractReader/V2ContractReader'
+import useContractReader from '../v2v3/contractReader/V2ContractReader'
 
 /** Returns ERC20 balance of `wallet`. Pass arguments for `projectId` and `terminalName` if the ERC20 is a project token, to update the returned value when relevant on-chain events are emitted. */
 export default function useERC20BalanceOf(
