+++ src/hooks/v1/V1ProjectState.ts
@@ -2,7 +2,7 @@ import { ProjectMetadataContext } from 'contexts/projectMetadataContext'
 import { V1ProjectContextType } from 'contexts/v1/projectContext'
 import { useCurrencyConverter } from 'hooks/CurrencyConverter'
 import { useProjectsQuery } from 'hooks/Projects'
-import useSymbolOfERC20 from 'hooks/SymbolOfERC20'
+import useSymbolOfERC20 from 'hooks/ERC20/SymbolOfERC20'
 import { V1CurrencyOption } from 'models/v1/currencyOption'
 import { useContext, useMemo } from 'react'
 import { V1CurrencyName } from 'utils/v1/currency'
