+++ src/hooks/v2v3/V2V3ProjectState.ts
@@ -7,9 +7,9 @@ import {
 import { V2V3ProjectContextType } from 'contexts/v2v3/V2V3ProjectContext'
 import { V2V3ProjectContractsContext } from 'contexts/v2v3/V2V3ProjectContractsContext'
 import { useCurrencyConverter } from 'hooks/CurrencyConverter'
-import useNameOfERC20 from 'hooks/NameOfERC20'
+import useNameOfERC20 from 'hooks/ERC20/NameOfERC20'
 import { useProjectsQuery } from 'hooks/Projects'
-import useSymbolOfERC20 from 'hooks/SymbolOfERC20'
+import useSymbolOfERC20 from 'hooks/ERC20/SymbolOfERC20'
 import { useBallotState } from 'hooks/v2v3/contractReader/BallotState'
 import { usePaymentTerminalBalance } from 'hooks/v2v3/contractReader/PaymentTerminalBalance'
 import useProjectCurrentFundingCycle from 'hooks/v2v3/contractReader/ProjectCurrentFundingCycle'
